ai-bid-system/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-docx>=0.8.11
python-multipart>=0.0.6
pydantic>=2.0.0
